requirements/base.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-environ==0.11.2
celery[redis]==5.4.0
django-celery-beat==2.7.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.3.0
